Move import_folders to new module
