+++ stream_alert/rules_engine/rules_engine.py
@@ -22,10 +22,11 @@ from stream_alert.rules_engine.threat_intel import ThreatIntel
 from stream_alert.shared import resources, RULES_ENGINE_FUNCTION_NAME as FUNCTION_NAME
 from stream_alert.shared.alert import Alert
 from stream_alert.shared.config import load_config
-from stream_alert.shared.rule import import_folders, Rule
+from stream_alert.shared.importer import import_folders
 from stream_alert.shared.logger import get_logger
 from stream_alert.shared.lookup_tables import LookupTables
 from stream_alert.shared.metrics import MetricLogger
+from stream_alert.shared.rule import Rule
 from stream_alert.shared.rule_table import RuleTable
 from stream_alert.shared.stats import get_rule_stats
